+++ apps/web/src/types/post.ts
@@ -41,7 +41,6 @@ export interface Post {
   uploadTime: string;
   displayOrder: number;
   isLoading?: boolean;
-  postGroupId: PostGroupId;
 }
 
 export interface CreatedPost {
